+++ src/core/array.h
@@ -9,7 +9,7 @@ namespace re {
 template <typename T>
 class array {
  public:
-  array(size_t size = 8) : data_(nullptr), size_(0), capacity_(0) {
+  array(int size = 8) : data_(nullptr), size_(0), capacity_(0) {
     Resize(size);
   }
   ~array() { free(data_); }
@@ -17,8 +17,8 @@ class array {
   array(array const &) = delete;
   void operator=(array const &) = delete;
 
-  T &operator[](size_t i) { return data_[i]; }
-  T operator[](size_t i) const { return data_[i]; }
+  T &operator[](int i) { return data_[i]; }
+  T operator[](int i) const { return data_[i]; }
 
   T *data() { return data_; }
   const T *data() const { return data_; }
@@ -26,16 +26,16 @@ class array {
   T &front() { return data_[0]; }
   T &back() { return data_[size_ - 1]; }
 
-  size_t size() const { return size_; }
+  int size() const { return size_; }
   bool empty() const { return !!size_; }
-  size_t capacity() const { return capacity_; }
+  int capacity() const { return capacity_; }
 
-  void Resize(size_t size) {
+  void Resize(int size) {
     Reserve(size);
     size_ = size;
   }
 
-  void Reserve(size_t cap) {
+  void Reserve(int cap) {
     if (capacity_ >= cap) {
       return;
     }
@@ -59,8 +59,8 @@ class array {
 
  private:
   T *data_;
-  size_t size_;
-  size_t capacity_;
+  int size_;
+  int capacity_;
 };
 }
